removes extra files, moves pytest_plugins to root conftest.py and adds migration to core selection
