+++ services/web/server/tests/conftest.py
@@ -1,12 +1,8 @@
-""" Main test configuration
-
-    EXPECTED: simcore_service_webserver installed
-
-"""
 # pylint: disable=unused-argument
 # pylint: disable=bare-except
 # pylint: disable=redefined-outer-name
 
+import json
 import logging
 import sys
 from pathlib import Path
@@ -14,11 +10,10 @@ from typing import Dict
 
 import pytest
 import simcore_service_webserver
-from integration.utils import get_fake_data_dir, get_fake_project
 from pytest_simcore.helpers.utils_login import LoggedUser, UserDict
 from simcore_service_webserver.db_models import UserRole
 
-current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
+CURRENT_DIR = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
 
 
 log = logging.getLogger(__name__)
@@ -26,6 +21,27 @@ log = logging.getLogger(__name__)
 # mute noisy loggers
 logging.getLogger("openapi_spec_validator").setLevel(logging.WARNING)
 logging.getLogger("sqlalchemy").setLevel(logging.WARNING)
+
+
+# imports the fixtures for the integration tests
+pytest_plugins = [
+    "pytest_simcore.celery_service",
+    "pytest_simcore.docker_compose",
+    "pytest_simcore.docker_registry",
+    "pytest_simcore.docker_swarm",
+    "pytest_simcore.environment_configs",
+    "pytest_simcore.monkeypatch_extra",
+    "pytest_simcore.postgres_service",
+    "pytest_simcore.pydantic_models",
+    "pytest_simcore.rabbit_service",
+    "pytest_simcore.redis_service",
+    "pytest_simcore.repository_paths",
+    "pytest_simcore.schemas",
+    "pytest_simcore.services_api_mocks_for_aiohttp_clients",
+    "pytest_simcore.simcore_services",
+    "pytest_simcore.tmp_path_extra",
+    "pytest_simcore.websocket_client",
+]
 
 
 @pytest.fixture(scope="session")
@@ -37,6 +53,14 @@ def package_dir() -> Path:
 
 
 @pytest.fixture(scope="session")
+def project_slug_dir(osparc_simcore_root_dir) -> Path:
+    service_folder = osparc_simcore_root_dir / "services" / "web" / "server"
+    assert service_folder.exists()
+    assert any(service_folder.glob("src/simcore_service_server"))
+    return service_folder
+
+
+@pytest.fixture(scope="session")
 def api_specs_dir(osparc_simcore_root_dir: Path) -> Path:
     specs_dir = osparc_simcore_root_dir / "api" / "specs" / "webserver"
     assert specs_dir.exists()
@@ -44,15 +68,17 @@ def api_specs_dir(osparc_simcore_root_dir: Path) -> Path:
 
 
 @pytest.fixture(scope="session")
-def fake_data_dir() -> Path:
-    fake_data_dir = get_fake_data_dir()
-    assert fake_data_dir.exists()
-    return fake_data_dir
+def fake_data_dir(project_tests_dir: Path) -> Path:
+    data_dir = project_tests_dir / "data"
+    assert data_dir.exists()
+    return data_dir
 
 
 @pytest.fixture
 def fake_project(fake_data_dir: Path) -> Dict:
-    return get_fake_project()
+    fpath = fake_data_dir / "fake-project.json"
+    assert fpath.exists()
+    return json.loads(fpath.read_text())
 
 
 @pytest.fixture()
