+++ services/web/server/tests/integration/01/test_project_workflow.py
@@ -43,6 +43,7 @@ API_VERSION = "v0"
 pytest_simcore_core_services_selection = [
     "catalog",
     "director",
+    "migration",
     "postgres",
     "redis",
 ]
