+++ services/web/server/tests/integration/02/test_rabbit.py
@@ -38,9 +38,10 @@ from tenacity.wait import wait_fixed
 
 # Selection of core and tool services started in this swarm fixture (integration)
 pytest_simcore_core_services_selection = [
+    "migration",
     "postgres",
-    "redis",
     "rabbit",
+    "redis",
 ]
 
 pytest_simcore_ops_services_selection = []
