+++ services/web/server/tests/unit/conftest.py
@@ -20,27 +20,15 @@ from pytest_simcore.helpers.utils_projects import empty_project_data
 from simcore_service_webserver.resources import resources
 from simcore_service_webserver.rest import get_openapi_specs_path, load_openapi_specs
 
-## current directory
-current_dir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
+CURRENT_DIR = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
 
-## Log
 log = logging.getLogger(__name__)
-
-pytest_plugins = [
-    "pytest_simcore.environment_configs",
-    "pytest_simcore.monkeypatch_extra",
-    "pytest_simcore.pydantic_models",
-    "pytest_simcore.repository_paths",
-    "pytest_simcore.schemas",
-    "pytest_simcore.services_api_mocks_for_aiohttp_clients",
-    "pytest_simcore.websocket_client",
-]
 
 
 @pytest.fixture(scope="session")
 def here() -> Path:
     cdir = Path(sys.argv[0] if __name__ == "__main__" else __file__).resolve().parent
-    assert cdir == current_dir, "Somebody changing current_dir?"
+    assert cdir == CURRENT_DIR, "Somebody changing current_dir?"
     return cdir
